API/requirements.txt
tensorflow==2.16.2
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.5.0
matplotlib
numpy